Modularize and test
