+++ testData/test.legend
@@ -1,0 +1,32 @@
+id	position	a0	a1
+19:14492336_A_G	1	A	G
+19:14492337_T_C	2	T	C
+19:14492338_G_A	3	G	A
+19:14492339_G_A	4	G	A
+19:14492340_G_A	5	G	A
+19:14492341_A_C	6	A	C
+19:14492342_G_A	7	G	A
+19:14492343_G_T	8	G	T
+19:14492344_C_T	9	C	T
+19:14492345_C_A	10	C	A
+19:14492346_G_A	11	G	A
+19:14492347_C_T	12	C	T
+19:14492348_G_A	13	G	A
+19:14492349_T_A	14	T	A
+19:14492350_C_T	15	C	T
+19:14492351_T_C	16	T	C
+19:14492352_T_C	17	T	C
+19:14492353_T_A	18	T	A
+19:14492354_C_T	19	C	T
+19:14492355_T_G	20	T	G
+19:14492350_C_T	21	C	T
+19:14492351_T_C	22	T	C
+19:14492352_T_C	23	T	C
+19:14492353_T_A	24	T	A
+19:14492354_C_T	25	C	T
+19:14492355_T_G	26	T	G
+19:14492350_C_T	27	C	T
+19:14492351_T_C	28	T	C
+19:14492352_T_C	29	T	C
+19:14492353_T_A	30	T	A
+19:14492354_C_T	31	C	T
