+++ testData/testBins.txt
@@ -1,0 +1,4 @@
+Lower	Upper	Expected_var
+1	1	15
+2	2	8
+3	5	2
